requirements.txt
Flask
google-genai
pytest
bandit
